Rewrote a method and improved behavior
Rewrote the File Manager Form's 'lstDirectory_DoubleClick' method so it
is more clear, to prevent unnecessary strings being recreated, and
because it was making the same decision twice when it didn't need to.
Also placed the Setting Form's 'btnListen_Click' method in a try-finally
block so if anything goes wrong in the method, the state of it won't be
frozen (before, you would have to click out of the Form, discard
changes, then re-open. Perhaps it would be advisable to notify the user
of an issue with the listening if it failed...