Reflected changes to Server's UnsafeStreamCodec
Removed unnecessary comments (documentation) because if we need it, we
can look at older versions instead of letting it clutter the current
code.
Instead of disposing bitmaps at end of code, placed in a using because
it makes sure that the bitmap will always be disposed of properly.
Throws more meaningful exceptions.