FrmMain instance null check
Added FrmMain instance null check to safely exit a thread in the
FrmKeylogger form in the case of client-server disconnection.  The
FrmKeylogger spawns a thread that waits for the enabled button to be set
to true, after successfully receiving all the log files from the client.
If the server were to somehow not process the button being enabled it
would hang in the while loop leaving the entire process to be left in
memory due to the thread being open.