Last of the big changes to the Keylogger
Added the second and likely last of the big changes and additions to the
Keylogger. The core design of it is much more clear than the initial
changes commit. Though not yet functional (last of it has to be
implemented, as seen by the small commented-out portions), it is nearing
completion. With this new design, it is significantly more scalable.
With these changes also comes another huge factor: modifications. This
new system allows easy feature additions and modifications, such as
filtering out specific keys to log dynamically in a natural manner.