Disposing IDisposable Resources
Dispose of unmanaged resources used by the server's UnsafeStreamCodec
and JpgCompression (because it is pretty related to the
UnsafeStreamCodec). Made sure that the UnsafeStreamCodec was being
disposed of before being re-assigned.
Note: I did not reflect these changes to the client's UnsafeStreamCodec
because the implementation of the client's codec is a bit different in
the CommandHandler. This different implementation caused instability in
the client. In the server, I noticed decreased memory consumption and
decreased fluctuation in the memory used.