Modifier keys + key press fix
-if any modifier keys are down and a character is pressed on the
keyboard, it will look for a Key from the key character, if it doesn't
find it, it will process the character normally.  This is as best as I
could do for Alt Gr because Alt Gr translates to (LControlKey |
ControlKey).  In order to get Alt Gr presses to work, normally you would
have to press (Ctrl + Alt + key) to write the characters such as the
ones here highlighted in blue:
http://en.wikipedia.org/wiki/German_keyboard_layout#/media/File:KB_Germany.svg