Fixed occasional inability to disconnect server
It is possible that the client can not disconnect properly. This is most
often because of an issue disposing the user's forms.  If something
causes the client to not be able to properly disconnect, don't worry
about it and just remove them from the list.
Also unsubscribe from these events to avoid memory leaking, and use an
eventLocker readonly object to lock events and event handlers.