Clone Bitmap on FrmRemoteDesktop
Clone the Bitmap on the "UpdateImage" method. It should be the
responsibility of "UpdateImage" anyways because it is the method
responsible for specifying how to manage the Bitmap provided. This also
serves an issue where the call to clone the Bitmap and cast it to a
Bitmap object was no longer in a try-catch block as it used to be (and
should be once more).