Fixed Serialization and other changes
Finally found the issue with serialization.
While the RegSeekerMatches could be serialized, the RegSeekerParams
(settings for the Registry Seeking class) could not be due to the
RegistryKey array of root keys to base the search off of.
These changes finally demonstrate proper searching and the ability to
add the Registry Keys from the Client to the Server. It needs work on
proper order and searching, as well as proper parsing, but this is the
start of it.